requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
openai>=2.26.0
google-generativeai>=0.7.0
httpx>=0.27.0
python-dotenv>=1.0.0
pytest>=7.4.0
